Fix some errors in processing Text and Arrow shapes
The "size" field might not always be sent on the first update of a text
shape, so ensure it gets initialized (I use 0, 0 here to indicate
undetermined size; special handling is added so that this results in no
word wrapping being performed.

The text shape parsing was attempting to update some attributs related
to cached paths, but this shape doesn't use cached paths. Drop that.

Add a few attributes that the arrow shapes were missing. It's far from
complete, currently hitting an arrow still crashes rendering.