Rework shape init to avoid conflicting __init__
attrs provides an auto-generated init, and it conflicts with the
manually-generated init. Instead of trying to override it, put back
the "from_data" method to create a shape from json data.

Unlike the previous implementation, this runs the __init__ method to
set default values before loading data, so missing data is handled
cleanly.