New tldraw whiteboard slide scaling calculation
Starting with BigBlueButton 2.6.0-rc.4, the slides are scaled to fit
in a fixed dimension whiteboard, rather than having the whiteboard size
be the same as the slide size. This fixes issues where the line widths
and sizes of objects like sticky notes would appear to vary dramatically
on different slides.

Adapt the shape scaling code to handle this change. A lot of the code
is shared with the previous BBB whiteboard code, since a similar scaling
method was used there.