Handle arrow updates which remove decorations
When drawing a line (arrow with no decorations), tldraw omits the keys
for the 'start' and 'end' decorations from the decorations object in the
data json completely, rather than leaving the keys and setting the value
to null.

Handle this by removing the decoration if it was omitted in the json.

Fixes #27