Fix horizontal text alignment in show_layout_by_lines
Since we're not using PangoCairo's show_layout function, we aren't
getting the horizontal positioning of the line done for us.

However, if we switch to using the Pango.layout_iter to iterate over
lines, we can ask Pango what horizontal position it was going to use for
the line, and use that along with our own vertical position.

Fixes #24