Handle arrow updates with empty handle hash.
For some reason when updating the position of a shape, tldraw sometimes
sends and update with the "start" handle present, but its value is an
empty hash (no "point").

Wrap the code in an exception handler for KeyError so I don't have to
individually check each level of the hash.

Fixes #36