Limit the max size of image surface to render PDF
In order to work around some problems with Poppler when rendering PDF
files clipped, where some content that's supposed to be on the page
might end up invisible or distorted, we render the entire page at the
selected zoom level to an image surface, then render that (with
clipping) to the final surface.

The problem is that with excessively high zoom levels, the image surface
for the full page might exceed cairo's image surface limits: 2^15-1 or
just use a large amount of memory.

Apply limits to the max size of surface used. The limits allow up to 10x
zoom or cairo's max image size, whichever is lower. If the page is
zoomed in to a level that would exceed that, the lower resolution PDF
render gets scaled up using blurry image scaling.

Fixes #40