Fix math domain error calculating ellipse perimeter
The random offsets applied to the radiuses of the ellipse in the
draw_stroke_points method can be large enough that a very narrow
ellipse will end up with a negative radius, which breaks the following
calculation to estimate the perimeter. Ensure that even with the
offsets, the radius never goes below 0.

Fixes #44