Ignore tldraw add shape with missing type field.
We sometimes see events produced by BigBlueButton which look like
"update" events - they only have an id and point field (what you'd see
if you drag a shape to a different position), but they refer to a shape
id that hasn't previously been seen in the events file.

There's nothing that can be done except ignore the shape (and log a
message).

My suspicion is that the presenter might have triggered a bug which let
them drag the slide background image, since the events for the
background image are filtered out of the events.