Add poll title, improve style somewhat
Poll now uses the colors from tldraw for the
background, which makes it look more like the
poll in the live meeting. Text size and padding
is adjusted slightly smaller, since the tldraw
shapes end up being kind of small for the amount
of text in them.

Poll title is rendered at the top of the box if
present.