Limit the maximum zoom level
The BigBlueButton client (in current versions) doesn't allow zooming in
further than 400% (i.e. 1/4 of document width). However, there are
occasionally glitches which can cause a higher zoom level to be
reported. Sometimes *extremely* high (an issue was reported where the
events.xml `<widthRatio>` was 0.06944444444444445 - that corresponds to
a zoom level of 144000%).

This can cause issues in the rendering. For example, if a shape with
text is being drawn, the required font size might exceed the maximum
limits of the freetype text rasterizer. (With newer versions of cairo,
this error is recoverable, but with older versions like used on Ubuntu
20.04 the error will crash the video generation tool.)

Apply a safety limit on the zoom level so that it does not exceed 800%
(i.e. 1/16 of document width).