Add support for raw video output
Rather than pass the video to ffmpeg for encoding to a file, support
directly outputting the raw video frames.

This is expected to be used by the BigBlueButton recording processing,
which is already using ffmpeg, so it is able to accept the raw frames
directly via a pipe. This will allow it to render portions of the video
on demand, meaning the presentation area size can change, and the
presentation video does not need to be generated for sections of video
where the presentation is not visible.